fix:유형 A뉴스 오류 수정완료
