+++ classify_news/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
-from .views import ClassifiesAPIView, ClassifyAPIView, ClassifyCAPIView
+from .views import ClassifiesAPIView, ClassifyCAPIView
 
 urlpatterns = [
     path("", ClassifiesAPIView.as_view()),
-    path("A/<int:news_id>", ClassifyAPIView.as_view()),
     path("C/<int:news_id>", ClassifyCAPIView.as_view()),
 ]
