+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: '3'
 services:
   #cadvisor:
   #  image: google/cadvisor:latest
@@ -58,7 +57,6 @@ services:
       "
     volumes:
       - .:/Backend
-      - ./service-account-file.b64:/Backend/service-account-file.b64
     depends_on:
       - db
     env_file:
