+++ news/news_content.txt
@@ -1,0 +1,1 @@
+♣id:00178,content:우주산업 클러스터 3각 체제 구축.(대전시 제공) ⓒ News1 김경훈 기자전남과 경남, 대전을 잇는 우주 삼각 클러스터가 본격 추진된다. 정부는 2030년까지 총 3808억 원의 예산을 투자해 세 도시를 우주 특화 도시로 길러내겠다는 방침이다.과학기술정보통신부는 24일 ‘제6회 국가연구개발사업평가 총괄위원회’를 개최하고 우주산업클러스터 삼각체제 구축 사업에 대한 사업계획 적정성 검토 결과를 심의·의결했다고 밝혔다. 전남은 발사체, 경남은 위성, 대전은 연구·인재 육성을 각각 담당할 예정이다.통과된 사업계획에 따르면 전남 발사체 특화지구에는 전남 고흥 나로우주센터 내 민간 우주발사체 발사를 위한 전용 발사장이 들어설 계획이다. 그간 해외 발사장을 이용했던 민간 기업들의 발사비용이 줄어들 전망이다.경남 위성 특화지구에는 민간이 개발한 위성을 우주 유사 환경에서 시험할 수 있도록 ‘우주환경시험시설’과  위성개발혁신센터를 구축한다.대전 한국과학기술원(KAIST)에는 우주기술혁신인재양성센터를 구축해 실무 중심의 교육과정을 개발·운영할 예정이다.정부는 지난해 8월 이 사업의 예비타당성 조사 면제를 결정할 당시에는 6000억 원 수준으로 사업 규모를 산정했지만 이번 적정성 검토 단계에서 다소 축소됐다. 우주항공청은 “이번 적정성 검토 결과에 따른 총사업비는 기반시설 위주로 반영된 것으로 연구개발(R&D), 인력양성 관련 사업은 추후 별도 사업으로 추진하겠다”고 밝혔다.
