+++ news/sentiment.py
