+++ news/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import news_APIView, news_list_APIView, CrawlNewsView, CountNewsAPIView, NewsTimelineAPIView
+from .views import news_APIView, news_list_APIView, CrawlNewsView, CountNewsAPIView, NewsTimelineAPIView, AAPIView
 
 urlpatterns = [
     path('', news_APIView.as_view()),
@@ -7,4 +7,5 @@ urlpatterns = [
     path('crawlnews',CrawlNewsView.as_view()),
     path('count', CountNewsAPIView.as_view()),
     path('timeline', NewsTimelineAPIView.as_view()),
+    path("A/<int:news_id>", AAPIView.as_view()),
 ]
